fix(gl): place keeper sprite frames on the same pixels as software

+++ src/engine_render.h
@@ -217,7 +217,7 @@ TbBool resolve_keepersprite_draw_data(unsigned short anim_sprite, short angle,
 // heap) directly.
 TbBool resolve_keepersprite_cursor_geometry(short x, short y, unsigned short kspr_base,
     short kspr_angle, unsigned char sprgroup, long scale,
-    int32_t *out_dst_x, int32_t *out_dst_y, int32_t *out_dst_w, int32_t *out_dst_h,
+    float *out_dst_x, float *out_dst_y, float *out_dst_w, float *out_dst_h,
     int32_t *out_draw_idx, const unsigned char **out_data, int *out_src_w, int *out_src_h);
 void draw_map_volume_box(long cor1_x, long cor1_y, long cor2_x, long cor2_y, long floor_height_z, unsigned char color);
