+++ src/kfx/renderer/backends/SoftwareWorldViewRenderer.h
@@ -25,7 +25,7 @@ public:
     void MarkDeferredWorldAsLensCapture() override { m_lens = true; }
 
     void SubmitKeeperSprite(int32_t frame_x, int32_t frame_y,
-                            int32_t dst_x, int32_t dst_y, int32_t dst_w, int32_t dst_h,
+                            float dst_x, float dst_y, float dst_w, float dst_h,
                             const unsigned char* data, int src_w, int src_h, int32_t content_h,
                             unsigned int draw_flags, const unsigned char* remap,
                             int32_t sprite_id) override;
