+++ src/kfx/renderer/ir/WorldCommands.h
@@ -80,10 +80,10 @@ struct IRWorldTexQuadCmd
 struct IRWorldKeeperSpriteCmd
 {
     WorldCmdLayer         layer         = WorldCmdLayer::Sprites;
-    int32_t               dst_x         = 0;   /**< Screen destination left. */
-    int32_t               dst_y         = 0;   /**< Screen destination top. */
-    int32_t               dst_w         = 0;   /**< Destination width. */
-    int32_t               dst_h         = 0;   /**< Destination height. */
+    float                 dst_x         = 0.0f; /**< Screen destination left (sub-pixel). */
+    float                 dst_y         = 0.0f; /**< Screen destination top (sub-pixel). */
+    float                 dst_w         = 0.0f; /**< Destination width. */
+    float                 dst_h         = 0.0f; /**< Destination height. */
     int32_t               src_w         = 0;   /**< Source sprite width. */
     int32_t               src_h         = 0;   /**< Source sprite height -- also the atlas decode/cache height (always the sprite's full, unclipped content, so a later full-height draw of the same sprite_id never samples missing rows). */
     /** Visible rows out of @p src_h for THIS draw (water/lava
